WIP. Refactor structure code. Working out bugs from StructureBuilder.
