+++ protein_transformer/models/rnn.py
@@ -5,7 +5,7 @@ import torch
 from torch import nn
 from torch.autograd import Variable
 
-from protein_transformer.protein.Sidechains import NUM_PREDICTED_ANGLES
+from protein_transformer.protein.Structure import NUM_PREDICTED_ANGLES
 
 
 class MyRNN(nn.Module):
